rewrite JapaneseTokenizer
